Fixed Bugs

+++ test.c
@@ -35,7 +35,7 @@ int testing() {
 }
 
 int testIsCorrect(const int i, const double a, const double b, const double c,
-                  const double rootsNum, const double dataX1, const double dataX2) {
+                  const int rootsNum, const double dataX1, const double dataX2) {
     double x1 = 0, x2 = 0;
     int answers = squareEquationRootsNum(a, b, c, &x1, &x2);
 
@@ -48,8 +48,8 @@ int testIsCorrect(const int i, const double a, const double b, const double c,
     return 0;
 }
 
-int testFailure (const int i, const double x1, const double x2, const double answers,
-                 const double rootsNum, const double dataX1, const double dataX2) {
+int testFailure (const int i, const double x1, const double x2, const int answers,
+                 const int rootsNum, const double dataX1, const double dataX2) {
     printf("Test number %d has been failed:\n\n", i);
     printf("Received answer:\n\n1st root: %lg\n2nd Root: %lg\nNumber of roots: %d\n\n", x1, x2, answers);
     printf("Right answer:\n\n1st root: %lg\n2nd Root: %lg\nNumber of roots: %d\n\n", dataX1, dataX2, rootsNum);
